working to pass in props correctly
